test: split provider-discovery tests by module #382
Move the three concerns of the original `test_provider_discovery.py` into
files mirroring the production layout: config validation, the proxy
data-source handler, and the SDMX 2.1 data-source handler.
