--- tests/unit/common/data/statgpt_sdmx_proxy/v30/test_provider_discovery.py
@@ -1,175 +1,0 @@
-"""Tests for provider-scoped dataset listing on the StatGPT SDMX proxy."""
-
-import json
-from pathlib import Path
-
-import httpx
-import pytest
-from pydantic import ValidationError
-from sdmx.message import StructureMessage
-from sdmx.model.common import Agency, AgencyScheme, InternationalString
-
-from statgpt.common.data.base.datasource import ProviderRequiredError
-from statgpt.common.data.sdmx.common.config import (
-    ProviderDiscoveryMode,
-    SdmxConfig,
-    SdmxDataSourceConfig,
-)
-from statgpt.common.data.sdmx.v21.datasource import Sdmx21DataSourceHandler
-from statgpt.common.data.statgpt_sdmx_proxy.config import StatGptSdmxProxyDataSourceConfig
-from statgpt.common.data.statgpt_sdmx_proxy.v30.datasource import StatGptSdmxProxyDataSourceHandler
-from statgpt.common.data.statgpt_sdmx_proxy.v30.sdmx_client import AsyncStatGptSdmxProxyClient
-from statgpt.common.schemas.data_source import Provider
-
-FIXTURE = Path(__file__).parent / "agency_schemes_response.json"
-STRUCTURE_CONTENT_TYPE = "application/vnd.sdmx.structure+json;version=2.0.0"
-
-
-def _proxy_config(**overrides) -> StatGptSdmxProxyDataSourceConfig:
-    return StatGptSdmxProxyDataSourceConfig(
-        sdmx_config=SdmxConfig(id="proxy", url="https://example.invalid", name="proxy"),
-        **overrides,
-    )
-
-
-def _sdmx_config(**overrides) -> SdmxDataSourceConfig:
-    return SdmxDataSourceConfig(
-        sdmx_config=SdmxConfig(id="sdmx", url="https://example.invalid", name="sdmx"),
-        **overrides,
-    )
-
-
-def test_proxy_config_rejects_dataflows_discovery_mode() -> None:
-    with pytest.raises(ValidationError) as exc:
-        _proxy_config(provider_discovery=ProviderDiscoveryMode.DATAFLOWS)
-    assert "provider_discovery='dataflows'" in str(exc.value)
-
-
-def test_proxy_config_accepts_agencyscheme_discovery_mode() -> None:
-    config = _proxy_config(provider_discovery=ProviderDiscoveryMode.AGENCYSCHEME)
-    assert config.provider_discovery is ProviderDiscoveryMode.AGENCYSCHEME
-
-
-def test_proxy_config_default_discovery_is_agencyscheme() -> None:
-    assert _proxy_config().provider_discovery is ProviderDiscoveryMode.AGENCYSCHEME
-
-
-async def test_proxy_list_datasets_requires_provider() -> None:
-    handler = StatGptSdmxProxyDataSourceHandler(_proxy_config())
-    with pytest.raises(ProviderRequiredError):
-        await handler.list_datasets(auth_context=None)  # type: ignore[arg-type]
-
-
-def test_proxy_agency_scheme_fixture_has_expected_shape() -> None:
-    """The fixture is a captured /structure/agencyscheme response. Keep it stable so the
-    parser tests stay anchored to the real proxy payload."""
-
-    payload = json.loads(FIXTURE.read_text())
-    schemes = payload["data"]["agencySchemes"]
-    assert schemes, "fixture should expose at least one agency scheme"
-    agencies = schemes[0]["agencies"]
-    ids = {agency["id"] for agency in agencies}
-    assert {"BIS", "IMF"}.issubset(ids)
-    by_id = {a["id"]: a["name"] for a in agencies}
-    assert by_id["BIS"] == "Bank for International Settlements"
-    assert by_id["IMF"] == "International Monetary Fund"
-
-
-def _agency_scheme_message(agencies: list[Agency]) -> StructureMessage:
-    msg = StructureMessage()
-    scheme = AgencyScheme(id="AGENCIES")
-    for agency in agencies:
-        scheme.items[agency.id] = agency
-    msg.add(scheme)
-    return msg
-
-
-class _StubClient:
-    def __init__(self, message: StructureMessage):
-        self._message = message
-
-    async def agencyscheme(self, *, agency_id, resource_id, version):
-        return self._message
-
-    async def dataflow(self, *, agency_id, resource_id, version, params=None, use_cache=False):
-        return self._message
-
-
-class _StubProxyHandler(StatGptSdmxProxyDataSourceHandler):
-    """Test double that returns a canned StructureMessage instead of hitting HTTP."""
-
-    def __init__(self, config: StatGptSdmxProxyDataSourceConfig, message: StructureMessage):
-        super().__init__(config)
-        self._stub_client = _StubClient(message)
-
-    async def create_sdmx_client(self, auth_context):  # type: ignore[override]
-        return self._stub_client  # type: ignore[return-value]
-
-
-def _mock_transport(body: bytes) -> httpx.MockTransport:
-    def respond(_: httpx.Request) -> httpx.Response:
-        return httpx.Response(200, headers={"content-type": STRUCTURE_CONTENT_TYPE}, content=body)
-
-    return httpx.MockTransport(respond)
-
-
-async def test_proxy_list_providers_parses_real_agency_scheme_response(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    transport = _mock_transport(FIXTURE.read_bytes())
-    monkeypatch.setattr(
-        AsyncStatGptSdmxProxyClient,
-        "_create_httpx_client",
-        staticmethod(lambda *, headers=None: httpx.AsyncClient(transport=transport)),
-    )
-
-    handler = StatGptSdmxProxyDataSourceHandler(_proxy_config())
-    providers = await handler.list_providers(auth_context=None)  # type: ignore[arg-type]
-
-    assert providers == [
-        Provider(id="BIS", name="Bank for International Settlements"),
-        Provider(id="IMF", name="International Monetary Fund"),
-    ]
-
-
-async def test_sdmx21_list_providers_falls_back_to_id_when_localized_name_missing() -> None:
-    agency = Agency(id="OECD")
-    agency.name = InternationalString()
-    message = _agency_scheme_message([agency])
-
-    handler = _StubProxyHandler(_proxy_config(), message)
-    providers = await handler.list_providers(auth_context=None)  # type: ignore[arg-type]
-
-    assert providers == [Provider(id="OECD", name="OECD")]
-
-
-class _StubSdmx21Handler(Sdmx21DataSourceHandler):
-    """Test double for the dataflows discovery mode."""
-
-    def __init__(self, config: SdmxDataSourceConfig, message: StructureMessage):
-        super().__init__(config)
-        self._stub_client = _StubClient(message)
-
-    async def create_sdmx_client(self, auth_context):  # type: ignore[override]
-        return self._stub_client  # type: ignore[return-value]
-
-
-async def test_sdmx21_list_providers_via_dataflows_dedupes_agency_ids() -> None:
-    from sdmx.model.v21 import DataflowDefinition
-
-    message = StructureMessage()
-    for agency_id, resource_id in [("IMF.RES", "WEO"), ("IMF.RES", "IFS"), ("BIS", "LBS")]:
-        dataflow = DataflowDefinition(
-            id=resource_id, maintainer=Agency(id=agency_id), version="1.0"
-        )
-        message.add(dataflow)
-
-    handler = _StubSdmx21Handler(
-        _sdmx_config(provider_discovery=ProviderDiscoveryMode.DATAFLOWS), message
-    )
-    providers = await handler.list_providers(auth_context=None)  # type: ignore[arg-type]
-
-    assert providers == [
-        Provider(id="BIS", name="BIS"),
-        Provider(id="IMF.RES", name="IMF.RES"),
-    ]
